apply comment from github&change field type uint64 to int64

+++ modules/farm/client/cli/tx.go
@@ -71,7 +71,7 @@ func GetCmdCreateFarmPool() *cobra.Command {
 				Name:           args[0],
 				Description:    description,
 				LpTokenDenom:   lpTokenDenom,
-				StartHeight:    uint64(startHeight),
+				StartHeight:    startHeight,
 				RewardPerBlock: rewardPerBlock,
 				TotalReward:    totalReward,
 				Editable:       editable,
